requirements.txt
streamlit
gspread
oauth2client
pandas
python-dotenv
serpapi
langchain
faiss-cpu
transformers
torch

